Adding scripts for recent experiments.

+++ scripts/train.tense.only.sh
@@ -4,14 +4,16 @@ bert_type=bert-base-cased
 seed=2222
 gec_model=../pseudo_model/ldc_giga.spell_error.pretrain.checkpoint_last.pt
 bert_model=../bert-base-cased
-experiment=early
-n_epochs=1
+experiment=tense-only
 checkpoint=checkpoint_pretrain
+epochs=5
+save_interval=10000
 
 SUBWORD_NMT=../subword
 FAIRSEQ_DIR=../bert-nmt
 BPE_MODEL_DIR=../gec-pseudodata/bpe
-DATA_DIR=~/data
+TRAIN_DIR=~/filter-data
+VALID_DIR=~/gec-data/conll14st-test-data/noalt
 VOCAB_DIR=../gec-pseudodata/vocab
 PROCESSED_DIR=../process/$experiment
 MODEL_DIR=../model/$bert_type/$experiment
@@ -19,12 +21,12 @@ LOG_DIR=../model/$bert_type/$experiment/logs
 
 pre_trained_model=../pretrained/ldc_giga.spell_error.pretrain.checkpoint_last.pt
 
-train_src=$DATA_DIR/dropna.1M.src
-train_trg=$DATA_DIR/dropna.1M.trg
-valid_src=$DATA_DIR/dropna.1K.src
-valid_trg=$DATA_DIR/dropna.1K.trg
-test_src=$DATA_DIR/dropna.1K.src
-test_trg=$DATA_DIR/dropna.1K.trg
+train_src=$TRAIN_DIR/bea.tense.only.src
+train_trg=$TRAIN_DIR/bea.tense.only.trg
+valid_src=$VALID_DIR/official-2014.combined.src 
+valid_trg=$VALID_DIR/official-2014.combined.trg
+test_src=$VALID_DIR/official-2014.combined.src
+test_trg=$VALID_DIR/official-2014.combined.trg
 
 cpu_num=`grep -c ^processor /proc/cpuinfo`
 
@@ -59,11 +61,14 @@ fi
 
 mkdir -p $MODEL_DIR
 
-cp $pre_trained_model $MODEL_DIR/checkpoint_pretrain.pt
+if [ ! -e "$MODEL_DIR/checkpoint_pretrain.pt" ]; then
+	echo "Copying pretrain file..."
+	cp $pre_trained_model $MODEL_DIR/checkpoint_pretrain.pt
+fi
 
-CUDA_VISIBLE_DEVICES=0 python3 -u $FAIRSEQ_DIR/train.py $PROCESSED_DIR/bin \
-    --save-dir $MODEL_DIR \
+CUDA_VISIBLE_DEVICES=0,1 python3 -u $FAIRSEQ_DIR/train.py $PROCESSED_DIR/bin \
     --tensorboard-logdir $LOG_DIR \
+    --save-dir $MODEL_DIR \
     --arch transformer_s2_vaswani_wmt_en_de_big \
     --max-tokens 4096 \
     --optimizer adam \
@@ -74,21 +79,21 @@ CUDA_VISIBLE_DEVICES=0 python3 -u $FAIRSEQ_DIR/train.py $PROCESSED_DIR/bin \
     --lr-scheduler reduce_lr_on_plateau \
     --lr-shrink 0.7 \
     --min-lr 1e-06 \
-    --restore-file $checkpoint.pt \
+    --warmup-from-nmt \
+    --warmup-nmt-file $checkpoint.pt \
     --bert-model-name $bert_model \
     --encoder-bert-dropout \
     --encoder-bert-dropout-ratio 0.3 \
     --clip-norm 1.0 \
     --criterion label_smoothed_cross_entropy \
     --label-smoothing 0.1 \
-    --max-epoch $n_epochs \
+    --max-epoch $epochs \
     --adam-betas '(0.9,0.98)' \
     --log-format simple \
-    --save-interval-updates 1000 \
+    --save-interval-updates $save_interval \
     --fp16 \
     --seed $seed \
     --reset-lr-scheduler \
     --reset-optimizer \
     --reset-meters \
     --reset-dataloader
-
